Did some casting to a float to actually pass around a float

+++ src/nmt/minecraft/Regicide/Game/GameTimer.java
@@ -36,7 +36,7 @@ public class GameTimer extends BukkitRunnable {
 			game.scorePoint();
 		}
 		
-		game.getBoard().update((time % scoreInterval) / scoreInterval);
+		game.getBoard().update(((float) (time % scoreInterval)) / (float) scoreInterval);
 		
 	}
